pyproject.toml
[project]
name = "tesislo"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "simpy>=4.1.1",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "streamlit>=1.28",
    "plotly>=5.18",
    "pulp>=2.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lemon_packing"]

[tool.uv]
dev-dependencies = ["pytest"]
